Add JUnit test for starting and closing the IntelliJ for UI tests, perform some refactoring
Fixes #17
