+++ gradle.properties
@@ -1,3 +1,3 @@
-projectVersion=0.0.1
+projectVersion=0.0.2-SNAPSHOT
 nexusUser=invalid
 nexusPassword=invalid
