+++ src/main/java/com/redhat/devtools/intellij/commonUiTestLibrary/UITestRunner.java
@@ -11,12 +11,22 @@
 package com.redhat.devtools.intellij.commonUiTestLibrary;
 
 import com.intellij.remoterobot.RemoteRobot;
-import com.redhat.devtools.intellij.commonUiTestLibrary.utils.GlobalUtils;
+import com.redhat.devtools.intellij.commonUiTestLibrary.fixtures.dialogs.WelcomeFrameDialog;
 
-import java.io.*;
+import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
+import java.io.FileOutputStream;
+import java.net.InetSocketAddress;
+import java.net.Socket;
+import java.net.SocketAddress;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.time.Duration;
+
+import static com.intellij.remoterobot.utils.RepeatUtilsKt.waitFor;
 
 /**
  * Basic methods for starting and quiting the IntelliJ Idea IDE for UI tests
@@ -34,20 +44,21 @@ public class UITestRunner {
         String osName = System.getProperty("os.name").toLowerCase();
         ProcessBuilder pb;
         if (osName.contains("windows")) {
-            pb = new ProcessBuilder("powershell.exe", "Start-Process", ".\\gradlew.bat", "-ArgumentList", "\"runIdeForUiTests -PideaVersion=" + ideaVersion + "\"", "-WindowStyle", "hidden");
+            pb = new ProcessBuilder(".\\gradlew.bat", "runIdeForUiTests", "-PideaVersion=" + ideaVersion, "-Drobot-server.port=" + port);
         } else {
             pb = new ProcessBuilder("./gradlew", "runIdeForUiTests", "-PideaVersion=" + ideaVersion, "-Drobot-server.port=" + port);
         }
 
         try {
             ideProcess = pb.start();
-            GlobalUtils.waitUntilIntelliJStarts(port);
-            robot = GlobalUtils.getRemoteRobotConnection(port);
+            waitUntilIntelliJStarts(port);
+            robot = getRemoteRobotConnection(port);
         } catch (IOException | InterruptedException e) {
             e.printStackTrace();
         }
 
-        GlobalUtils.clearTheWorkspace();
+        WelcomeFrameDialog wfd = robot.find(WelcomeFrameDialog.class, Duration.ofSeconds(10));
+        wfd.clearTheWorkspace();
         return robot;
     }
 
@@ -56,7 +67,11 @@ public class UITestRunner {
     }
 
     public static void closeIde() {
-        ideProcess.destroy();
+        if (robot.isWin()) {
+            robot.find(WelcomeFrameDialog.class, Duration.ofSeconds(10)).windowsCloseButton().click();
+        } else {
+            ideProcess.destroy();
+        }
     }
 
     public static RemoteRobot getRemoteRobotInstance() {
@@ -76,8 +91,7 @@ public class UITestRunner {
             String acceptedDir = System.getProperty("user.home") + "/.local/share/JetBrains/consentOptions";
             createDirectoryHierarchy(acceptedDir);
             copyFileFromJarResourceDir(acceptedSourceLocation, acceptedDir + "/accepted");
-        }
-        else if (osName.contains("os x")) {
+        } else if (osName.contains("os x")) {
             String plistSourceLocation = "com.apple.java.util.prefs.plist";
             String plistDir = System.getProperty("user.home") + "/Library/Preferences";
             copyFileFromJarResourceDir(plistSourceLocation, plistDir + "/com.apple.java.util.prefs.plist");
@@ -95,8 +109,7 @@ public class UITestRunner {
             } catch (IOException | InterruptedException e) {
                 e.printStackTrace();
             }
-        }
-        else if (osName.contains("windows")) {
+        } else if (osName.contains("windows")) {
             String acceptedSourceLocation = "accepted";
             String acceptedDir = System.getProperty("user.home") + "\\AppData\\Roaming\\JetBrains\\consentOptions";
             createDirectoryHierarchy(acceptedDir);
@@ -104,7 +117,7 @@ public class UITestRunner {
 
             String registryPath = "HKCU:\\Software\\JavaSoft\\Prefs\\jetbrains\\privacy_policy";
             ProcessBuilder pb1 = new ProcessBuilder("powershell.exe", "New-Item", "-Path", registryPath, "-Force");
-            ProcessBuilder pb2 = new ProcessBuilder("powershell.exe", "New-ItemProperty", "-Path", registryPath, "-Name", "accepted_version", "-Value", "\"2.1\"");
+            ProcessBuilder pb2 = new ProcessBuilder("powershell.exe", "New-ItemProperty", "-Path", registryPath, "-Name", "accepted_version", "-Value", "'2.1'");
 
             try {
                 Process p1 = pb1.start();
@@ -115,6 +128,39 @@ public class UITestRunner {
                 e.printStackTrace();
             }
         }
+    }
+
+    private static void waitUntilIntelliJStarts(int port) {
+        waitFor(Duration.ofSeconds(600), Duration.ofSeconds(3), "The IntelliJ Idea did not start in 10 minutes.", () -> isIntelliJUIVisible(port));
+    }
+
+    private static boolean isIntelliJUIVisible(int port) {
+        return isHostOnIpAndPortAccessible("127.0.0.1", port);
+    }
+
+    private static boolean isHostOnIpAndPortAccessible(String ip, int port) {
+        SocketAddress sockaddr = new InetSocketAddress(ip, port);
+        Socket socket = new Socket();
+
+        try {
+            socket.connect(sockaddr, 10000);
+        } catch (IOException IOException) {
+            return false;
+        }
+        return true;
+    }
+
+    public static RemoteRobot getRemoteRobotConnection(int port) throws InterruptedException {
+        RemoteRobot remoteRobot = new RemoteRobot("http://127.0.0.1:" + port);
+        for (int i = 0; i < 60; i++) {
+            try {
+                remoteRobot.find(WelcomeFrameDialog.class);
+            } catch (Exception ex) {
+                Thread.sleep(1000);
+            }
+        }
+
+        return remoteRobot;
     }
 
     private static void createDirectoryHierarchy(String location) {
